+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Logger.
 - GetDeleteByPrimaryKeyQuery, GetSelectByPrimaryKeyQuery and GetUpdateByPrimaryKeyQuery methods on ISqlProvider.
 - Extension methods evolving NPGSQL classes.
+- JWE authentication.
+- Methods to enable Swagger using v1.
 ### Removed
 - ReferencedPropertyName property from ForeignKeyColumnAttribute.